KeyWordSpotting/reset_se/TrainingProgram/requirement_for_quant.txt
absl-py==1.4.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
charset-normalizer==3.1.0
coloredlogs==15.0.1
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.2.1
executing==2.0.1
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.62.2
h5py==3.11.0
humanfriendly==10.0
importlib_metadata==7.1.0
ipython==8.18.1
jedi==0.19.1
keras==3.2.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.3.2
namex==0.0.8
oauthlib==3.2.2
onnx==1.16.0
onnx-graphsurgeon==0.5.2
onnx2tf==1.20.0
onnxruntime==1.17.3
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.17.2
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
six==1.16.0
sng4onnx==1.0.1
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-model-optimization==0.8.0
termcolor==2.4.0
tf_keras==2.16.0
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.18.1
